requirements.txt
fastapi
uvicorn
hashids
pyyaml
colorlog
orator
pydantic
psycopg2